+++ src/components/Calendar/CalendarInfo.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import styled from "styled-components";
+
+const CalendarInfo = () => {
+  return <Container></Container>;
+};
+
+// 임시 구획 확인용
+const Container = styled.section`
+  height: 100px;
+  background-color: gray;
+  border: 2px solid gray;
+`;
+
+export default CalendarInfo;
