+++ src/components/Calendar/Date.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Date = ({ children }) => {
+  return <div>{children}</div>;
+};
+
+export default Date;
